requirements.txt
pandas==1.5.3
boto3==1.26.89
requests==2.28.2
PyYAML==6.0 